Add instructions for postgresql installation, prefer background services for redis/rabbitmq over honcho
The local dev setup fails with fixtures when run with SQLite, but works
fine with postgresql. This could be due to some constraints that postgres
enforces but not sqlite. It's also better to use postgres for local dev
too because that would be consistent with production environment.
The other change is preferring background services for redis/rabbitmq
over starting them through honcho. This is because starting/stopping them
takes longer, especially for rabbitmq. Honcho should be quick to start and
stop. This also makes sure that task results persist in Redis and unprocessed
tasks persist in RabbitMQ over honcho restarts.