Implement pricing for Eleven Labs
One side effect of this change -- namely of summing up the cost
of Lipsync and TTS models in the LipsyncTTS recipe is that now
other LipsyncTTS are costlier too. Earlier, LipsyncTTS didn't
consider the cost of TTS (which seemed to be a flat 5 credits
for all TTS models).
This change will need a pricing review and might be modified to
only do this summing up for ElevenLabs TTS.